build.sh: Prevent user from deleting meta-rz-features
If PROP_DIR is given the value of ${WORK_DIR}/meta-rz-features when
running the build script the meta-rz-features directory would be
deleted without replacement, leading to a failed build.

+++ scripts/build.sh
@@ -260,7 +260,7 @@ install_prop_libs () {
 	echo "#################################################################"
 	echo "Installing proprietary libraries..."
 
-	if $PROP_LIBS_EXTRACTED; then
+	if [ $PROP_LIBS_EXTRACTED ] && [ ${PROP_DIR} != "${WORK_DIR}/meta-rz-features" ]; then
 		rm -rf ${WORK_DIR}/meta-rz-features
 		cp -r ${PROP_DIR} ${WORK_DIR}/meta-rz-features
 	elif [ ${FAMILY} == "rzg2" ]; then
